On MOSFETmodelling
I found an article on IEEE about MOSFET modelling on simulink. I learnt
that an S function needs to be created for modelling the output current
and input Vgs-Vds voltages using the basic MOSFET current equations. I
searched on creating an S function first 1 level then 2 level with tlc.
I tried to write the MATLAB code according to the prescribed template.
However, I am stuck at a point. where I can not put 2 variables into the
function and where I can not find those blue boxes to indicate the
location of the function m file as in the  examples provided in Simulink
Browser.