:smile:
I made some tuning on the parameters to see how they behave and wrote a
document (short one) in latex and added a zip file including the
documents.(I use sharelatex and unfortunately adding the doc in GitHub
requires an additional payment :smile: ) and another document which was
published by a student and normal member of IEEE. I guess it is a
student research? Anyhow, I used it to get a general idea for the range
of the parasitic parameters. All these are explained in the report I
added.