A topology final version
The topology is changed accordingly. There are 4 functions for necessary computations of changing parameters. The start point (n1) calculation method remains unknown.
The on-off times will not happen immediately, probably it will take one to three Ts points. Hence I wrote the code like that, comparing 3 values in the future or in the past.